pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiofiles==23.2.1",
    "aiohttp==3.9.1",
    "aioredis==2.0.1",
    "aiosmtplib==3.0.1",
    "bcrypt==4.1.2",
    "boto3==1.34.13",
    "botocore==1.34.13",
    "celery[redis]==5.3.4",
    "cryptography==41.0.7",
    "email-validator==2.1.0",
    "fastapi==0.104.1",
    "flower==2.0.1",
    "hiredis==2.2.3",
    "httpx==0.25.2",
    "jinja2==3.1.2",
    "kombu==5.3.4",
    "motor==3.3.2",
    "openpyxl==3.1.2",
    "pandas==2.1.4",
    "passlib[bcrypt]==1.7.4",
    "psutil==5.9.6",
    "pydantic==2.4.2",
    "pydantic-settings==2.0.3",
    "pymongo==4.5.0",
    "python-dateutil==2.8.2",
    "python-dotenv==1.0.0",
    "python-jose[cryptography]==3.3.0",
    "python-multipart==0.0.6",
    "pytz>=2026.1.post1",
    "redis>=4.6.0",
    "sentry-sdk==1.39.2",
    "uvicorn[standard]==0.24.0",
]
